imagery import fixes
