+++ frontend/src/portal-imagery/components/ImageryCard.tsx
@@ -1,4 +1,4 @@
-import { IImageryProject } from "../imageryProjects";
+import { IImageryProject } from "../types";
 import placeholder from "../../assets/images/placeholder.png";
 
 interface ImageryCardProps {
